+++ design-system/select/src/components/__tests__/unit/server-side-hydrate.tsx
@@ -10,7 +10,7 @@ afterEach(() => {
   jest.resetAllMocks();
 });
 
-test('should ssr then hydrate select correctly', async () => {
+test.skip('should ssr then hydrate select correctly', async () => {
   const [example] = await getExamplesFor('select');
   const Example = require(example.filePath).default; // eslint-disable-line import/no-dynamic-require